# Conflicts:
#	src/main/java/org/generation/italy/Main.java
#	src/main/java/org/generation/italy/codeSchool/model/data/implementations/InMemoryCourseRepository.java
#	src/main/java/org/generation/italy/codeSchool/model/entities/Attendance.java
#	src/main/java/org/generation/italy/codeSchool/model/entities/ClassRoom.java
#	src/main/java/org/generation/italy/codeSchool/model/entities/Course.java
#	src/main/java/org/generation/italy/codeSchool/model/entities/CourseEdition.java
#	src/main/java/org/generation/italy/codeSchool/model/entities/EditionModule.java
#	src/main/java/org/generation/italy/codeSchool/model/entities/Enrollment.java
#	src/main/java/org/generation/italy/codeSchool/model/entities/Level.java
#	src/main/java/org/generation/italy/codeSchool/model/entities/Person.java
#	src/main/java/org/generation/italy/codeSchool/model/entities/RemotePlatform.java
#	src/main/java/org/generation/italy/codeSchool/model/entities/Skill.java
#	src/main/java/org/generation/italy/codeSchool/model/entities/Student.java
#	src/main/java/org/generation/italy/codeSchool/model/entities/Teacher.java
#	src/main/java/org/generation/italy/codeSchool/model/services/implementations/StandardDidacticService.java
#	src/main/java/org/generation/italy/codeSchool/model/services/implementations/TestDidacticService.java
#	src/main/java/org/generation/italy/codeSchool/view/UserInterfaceConsole.java
#	src/test/java/org/generation/italy/codeSchool/model/services/implementations/StandardDidacticServiceTest.java

+++ src/main/java/org/generation/italy/codeSchool/model/entities/Enrollment.java
@@ -9,9 +9,8 @@ public class Enrollment {
     private Student student;
     private CourseEdition courseEdition;
     private String courseEvaluation;
-    private double courseVote;
+    private int courseVote;
     private String studentEvaluation;
     private double studentVote;
     private boolean hasPaid;
-
 }
